move twitterIdForFollowingList to .env
Given how hard Twitter makes it to find your Twitter account
ID, I figure it's sensitive-enough information that we might
as well hide it in the .env file as well.

I also found that the .env.example doesn't work when sourced
in Bash since it doesn't export the variables.  I _think_ it
is what's intended, though, so adding it in there.